+++ apps/mission/src/components/mission/dappStore/ecosystem/Container.tsx
@@ -1,7 +1,6 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/apps/blob/main/LICENSE)
 
-import Link from "next/link";
 import { Button } from "../Button";
 import { EcosystemCard } from "./Card";
 import { EcosystemProps, ecosystemData } from "./ecosystemData";
@@ -12,10 +11,11 @@ import {
   CLICK_ON_VIEW_ALL_DAPPS,
   CLCIK_ON_APPLY_TO_BE_PART_OF_THE_ECOSYSTEM,
 } from "tracker";
+import { PrimaryLink } from "ui-helpers";
 export const EcosystemContainer = () => {
   const { handlePreClickAction } = useTracker(CLICK_ON_VIEW_ALL_DAPPS);
   const { handlePreClickAction: clickApply } = useTracker(
-    CLCIK_ON_APPLY_TO_BE_PART_OF_THE_ECOSYSTEM,
+    CLCIK_ON_APPLY_TO_BE_PART_OF_THE_ECOSYSTEM
   );
   const handleViewAlldApps = () => {
     handlePreClickAction();
@@ -40,15 +40,14 @@ export const EcosystemContainer = () => {
           </h1>
           <h2 className="text-grayOpacity2">
             {t("dappStore.ecosystem.description")}
-            <Link
-              className="ml-2 text-red transition-all duration-200 ease-in-out hover:text-[#f26850] active:text-[#f0735d]"
+            <PrimaryLink
+              className="ml-2"
               href={GOOGLE_FORM_URL}
-              rel="noopener noreferrer"
-              target="_blank"
               onClick={handleGoogleFormClick}
+              aria-label="google form"
             >
               {t("dappStore.ecosystem.description2")}
-            </Link>
+            </PrimaryLink>
           </h2>
         </div>
         <Button